requirements.txt
diffusers>=0.30.3
accelerate>=0.34.2
transformers>=4.44.2
numpy==1.26.0
torch>=2.4.0
torchvision>=0.19.0
sentencepiece>=0.2.0

gradio>=4.44.0
imageio>=2.35.1
imageio-ffmpeg>=0.5.1
openai>=1.45.0
moviepy>=1.0.3
pillow==9.5.0
scikit-video
